requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-crispy-forms==2.4
django-debug-toolbar==5.2.0
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.9.0
xhtml2pdf==0.2.15